ref: Separate implemenation for processes
