+++ src/main/index.ts
@@ -1,0 +1,2 @@
+export { MainFrontend } from './frontend';
+export { MainBackend } from './backend';
