+++ src/renderer/index.ts
@@ -1,0 +1,2 @@
+export { RendererBackend } from './backend';
+export { RendererFrontend } from './frontend';
